Moved generateNewProjectName into ProjectDetailsHelpers.
