+++ __tests__/ProjectValidationActions.test.js
@@ -111,84 +111,3 @@ describe('ProjectValidationActions.updateProjectFolderToNameSpecification', () =
     await expect(store.dispatch(ProjectValidationActions.updateProjectFolderToNameSpecification())).rejects.toMatchSnapshot();
   });
 });
-
-describe('ProjectValidationActions.generateNewProjectName', () => {
-  const base_manifest = {
-      target_language: {
-        id: 'fr',
-        name: 'francais',
-        direction: 'ltr'
-      },
-      project: {
-        id: 'eph',
-        name: 'Ephesians',
-        resourceId: 'ult',
-        nickName: 'unfoldingWord Literal Text'
-      },
-      type: {
-        id: 'bible'
-      }
-    };
-
-  test('generate new project name', () => {
-    // given
-    const manifest = JSON.parse(JSON.stringify(base_manifest));
-    const expectedProjectName = 'fr_ult_eph_bible';
-
-    // when
-    const projectName = ProjectValidationActions.generateNewProjectName(manifest);
-
-    //then
-    expect(projectName).toEqual(expectedProjectName);
-  });
-
-  test('generate new project name lowercase', () => {
-    // given
-    const manifest = JSON.parse(JSON.stringify(base_manifest));
-    manifest.project.resourceId = "ULT";
-    const expectedProjectName = 'fr_ult_eph_bible';
-
-    // when
-    const projectName = ProjectValidationActions.generateNewProjectName(manifest);
-
-    //then
-    expect(projectName).toEqual(expectedProjectName);
-  });
-
-  test('generate new project name without resource id', () => {
-    // given
-    const manifest = JSON.parse(JSON.stringify(base_manifest));
-    delete manifest.project.resourceId;
-    const expectedProjectName = 'fr_eph_bible';
-
-    // when
-    const projectName = ProjectValidationActions.generateNewProjectName(manifest);
-
-    //then
-    expect(projectName).toEqual(expectedProjectName);
-  });
-
-  test('generate new project name without type', () => {
-    // given
-    const manifest = JSON.parse(JSON.stringify(base_manifest));
-    delete manifest.type.id;
-    const expectedProjectName = 'fr_ult_eph_text';
-
-    // when
-    const projectName = ProjectValidationActions.generateNewProjectName(manifest);
-
-    //then
-    expect(projectName).toEqual(expectedProjectName);
-  });
-
-  test('should not crash on empty manifest', () => {
-    // given
-    const manifest = {};
-
-    // when
-    const projectName = ProjectValidationActions.generateNewProjectName(manifest);
-
-    //then
-    expect(typeof projectName).toEqual('string');
-  });
-});
